+++ src/hooks/hydrated-effect.ts
@@ -1,12 +1,40 @@
-import { DependencyList, EffectCallback } from 'react'
+import { DependencyList, EffectCallback, useEffect } from 'react'
 import { useUpdateEffect } from './update-effect'
 
-export function useHydratedEffect(
+/**
+ * Hook that runs the effect only after the component has hydrated and first rendered.
+ *
+ * @param effect The effect callback to run.
+ * @param deps The dependency list for the effect.
+ * @param hasHydrated Boolean indicating if the component has hydrated.
+ */
+export function useHydratedUpdateEffect(
   effect: EffectCallback,
   deps: DependencyList,
   hasHydrated: boolean
 ) {
   useUpdateEffect(() => {
+    if (!hasHydrated) {
+      return
+    }
+
+    return effect()
+  }, [hasHydrated, ...deps])
+}
+
+/**
+ * Hook that runs the effect only after the component has hydrated.
+ *
+ * @param effect The effect callback to run.
+ * @param deps The dependency list for the effect.
+ * @param hasHydrated Boolean indicating if the component has hydrated.
+ */
+export function useHydratedEffect(
+  effect: EffectCallback,
+  deps: DependencyList,
+  hasHydrated: boolean
+) {
+  useEffect(() => {
     if (!hasHydrated) {
       return
     }