+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Added zoom callback to get rid of need for hydrated listeners.
 - Added hydration checks to stores.
 - Added hydrated effect for loading stores without sync issues.
 - Simplified sidebars to stop context problems on hot loading.